Backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-core
langgraph
pinecone-client
supabase
pydantic
python-jose[cryptography]
passlib[bcrypt]
aiofiles
pypdf
python-docx
openpyxl
tiktoken
numpy
pandas
sentence-transformers
torch
transformers 